pyproject.toml
[project]
name = "neurosaber-certificaciones-backend"
version = "0.1.0"
description = "Neurosaber web application for managing certifications."
authors = [
    {name = "rocha-p",email = "rochapedroignacio@gmail.com"},
    {name = "leoasson",email = "leoasson@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
package-mode = false
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "typing-extensions (>=4.14.0,<5.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "testcontainers (>=4.10.0,<5.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "injector (>=0.22.0,<0.23.0)",
    "fastapi-injector (>=0.8.0,<0.9.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "fastapi-filter (>=2.0.1,<3.0.0)",
    "fastapi-pagination (>=0.13.2,<0.14.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "boto3 (>=1.38.43,<2.0.0)",
    "sentry-sdk[fastapi] (>=2.31.0,<3.0.0)",
    "jinja2 (>=3.1.2,<4.0.0)",
    "reportlab (>=4.0,<5.0.0)",
    "pypdf (>=4.0,<5.0.0)",
    "qrcode[pil] (>=7.4,<8.0.0)",
    "Pillow (>=10.0,<11.0.0)",
    "apscheduler (>=3.10,<4.0.0)",
    "pytest-cov (>=6.0,<7.0)",
    "factory-boy (>=3.3,<4.0)",
    "faker (>=33.0,<40.0)",
    "respx (>=0.22,<1.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "app" }
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
